flaskApp/requirements.txt
Flask==1.0.2
requests==2.21.0
Flask_PyMongo==2.3.0
Flask_Cors==3.0.8
python_dateutil==2.8.1
gunicorn
dnspython